==> notebooks/exploration/part01_data_splitting.ipynb <==
# Part01 Data Splitting

# import modules
from typing import List, Dict, NamedTuple, Any

import pandas as pd
from IPython.display import display
import joblib
import yaml
import numpy as np
from sklearn.model_selection import train_test_split
from sklearn.preprocessing import MinMaxScaler

from src.classes.TrData import TrData
from src.commons.Utils import cleanse_transform_data

raw_data: pd.DataFrame = pd.read_csv(
    "../../dataset/00_raw/teleco_customer_churn.csv"
)

raw_data.head(n=5)

We will not do `Data Exploration` in this notebook. 

Data Exploration has been done in `part00_data_exploration.ipynb` notebook

In this notebook, we will focus on:
* splitting the data into training, validation, and testing set
* calculate mean for `TotalCharge` column from training dataset
* Impute the `null` in `TotalCharge` column with mean
* Save the calculated mean, so that it can be used again in the inference step

churn_label: pd.DataFrame 
df_ohe: pd.DataFrame

_ ,churn_label, df_ohe = cleanse_transform_data(rw_data=raw_data)

print(churn_label.dtypes)
print("--"*6)
print(df_ohe.dtypes)

Let's split the `df_ohe` and `churn_label` into train, validation, and test dataset

ohe_train: pd.DataFrame
ohe_val: pd.DataFrame
ohe_test: pd.DataFrame
churn_train: pd.DataFrame
churn_val: pd.DataFrame
churn_test: pd.DataFrame

ohe_train, ohe_test, churn_train, churn_test = train_test_split(
    df_ohe, 
    churn_label, 
    test_size=0.2, 
    shuffle=True, 
    random_state=200
)

ohe_train, ohe_val, churn_train, churn_val = train_test_split(
    ohe_train,
    churn_train, 
    test_size=0.25,
    shuffle=True, 
    random_state=200
)

print("--"*6)
print(f"ohe_train {ohe_train.shape} | churn_train {churn_train.shape}")
print("--"*6)
print(f"ohe_val {ohe_val.shape} | churn_val {churn_val.shape}")
print("--"*6)
print(f"ohe_val {ohe_test.shape} | churn_val {churn_test.shape}")

dataset_to_save: Dict[str, pd.DataFrame] = {
    "ohe_train": ohe_train, 
    "churn_train": churn_train,
    "ohe_val": ohe_val, 
    "churn_val": churn_val,
    "ohe_test": ohe_test,
    "churn_test": churn_test
}

# save split dataset
for each_key, each_df in dataset_to_save.items():
    each_df.to_csv(
        path_or_buf=f"../../dataset/01_interim/{each_key}.csv", 
        index=False
    )

Cool, so now we already have `train` dataset, `validation` dataset, and `train` dataset

Now, remember from previous notebook `part00_data_exploration`, we obtained 11 `null` in `TotalCharge` column once it was converted to "int64"

Now that we have split the dataset into `train` , `validation` and `test` dataset; we can calculate the mean from the `train` dataset, and then impute all the `null` in all dataset with the calculated mean from the training

We will save this calculated mean into `config.yaml` because we have to impute the same value to future `null` in future inferences

mean_total_charges: float = round(ohe_train["TotalCharges"].mean(), 2) 

# Save calculated mean_total_charges to YAML file for future uses
with open("../../config.yaml", "w") as file_stream:
    yaml.safe_dump(
        data={"MEAN_TOTAL_CHARGES": int(mean_total_charges)},
        stream=file_stream
    )

# impute null values with calculated mean_total_charges
ohe_train.fillna(value=mean_total_charges, inplace=True)

Let's find the range of each column in `ohe_train`

df_range: pd.DataFrame =  ohe_train.agg(func=["min", "max"])

display(df_range)

For those columns whose `max > 1` : we need to scale them so that they are in the same range as other columns

# Find list of columns whose  `Max` is greater than `1`
columns_to_scale: List[str] = df_range.columns[df_range.loc["max"] > 1].to_list() 

mm_scaler: MinMaxScaler = MinMaxScaler(
    feature_range=(0, 1)
)

# train scaler using train dataset
for each_cols in columns_to_scale:
    ds: pd.Series = ohe_train[[each_cols]].copy()
    ds_scaled = mm_scaler.fit_transform(ds)
    joblib.dump(
        value=mm_scaler,
        filename=f"../../models/scaler/mm_scaler_{each_cols}.pkl"
    ) # -->> save each scaler to pickle
    ohe_train[[each_cols]] = ds_scaled
    

display(ohe_train)

Alright, once we scale all numerical form to between `0` to `1`, we can convert the dataframe to numpy array for further processing later

# converting pandas dataframe to numpy

ohe_np: np.ndarray = ohe_train.to_numpy()
churn_np: np.ndarray = churn_train.to_numpy()

print(f"ohe_np shape {ohe_np.shape}")
print(f"churn_np shape {churn_np.shape}")

Hence, to recap, for the data_splitting, we are doing the following: 
1. impute `null` in `TotalCharges` with mean value
2. Scale `tenure` , `MonthlyCharge`, and `TotalCharges` using MinMaxScaler
3. Convert pandas dataframe to numpy

As usual, let's create a function to encapsulate all these steps, so that it can be reused

class NpData(NamedTuple):
    ohe_np: np.ndarray
    churn_np: np.ndarray

def impute_scale_and_convert_to_numpy(
        ohe_df: pd.DataFrame,
        columns_with_nulls: List[str],
        impute_val: List, 
        scaler_folder: str,
        churn_df: pd.DataFrame
) -> NpData:
    """
    
    :param ohe_df: dataframe to impute and scale 
    :param columns_with_nulls:  the column which has null values
    :param impute_val:  list of values to replace the nulls
    :param scaler_folder:  folder where we put the scaler
    :param churn_df: dataframe of the churn target
    :return: A tuple of scaled feature numpy array and churn target numpy array
    """
    
    # Impute null values
    for each_cols, im_val in zip(columns_with_nulls, impute_val):
        ohe_df[[each_cols]] = ohe_df[[each_cols]].fillna(value=im_val, inplace=False)
    
    # scale columns to between 0 and 1
    df_range: pd.DataFrame = ohe_df.agg(func=["min", "max"])
    columns_to_scale: List[str] = df_range.columns[df_range.loc["max"] > 1].to_list()
    
    for each_cols in columns_to_scale:
        ds: pd.Series = ohe_df[[each_cols]].copy()
        
        file_path: str = f"{scaler_folder}/mm_scaler_{each_cols}.pkl" 
        mm_scaler: MinMaxScaler = joblib.load(filename=file_path)
        ds_scaled = mm_scaler.transform(ds)
        ohe_df[[each_cols]] = ds_scaled
    
    ohe_np: np.ndarray = ohe_df.to_numpy()
    churn_np: np.ndarray = churn_df.to_numpy()
    
    return NpData(
        ohe_np=ohe_np,
        churn_np=churn_np
    )

Let's try our newly minted function 

_ohe_np: np.ndarray
_churn_np: np.ndarray

_ohe_np, _churn_np = impute_scale_and_convert_to_numpy(
    ohe_df=ohe_train,
    columns_with_nulls=["TotalCharges"],
    impute_val=[mean_total_charges],
    scaler_folder="../../models/scaler",
    churn_df=churn_train
)

print(f"_ohe_np shape {_ohe_np.shape}")
print("---"*6)
print(f"_churn_np shape {_churn_np.shape}")
